pyproject.toml
[build-system]
requires = ["setuptools>=75", "setuptools-scm>=8.1"]
build-backend = "setuptools.build_meta"

[project]
name = "mpeg1fs"
description = "A FUSE FS that transcodes to MPEG1"
license = {text = "AGPL-3.0-or-later"}
readme = "README.md"
dynamic = ["version"]
requires-python = ">= 3.10"
dependencies = [
    "fusepy>=3.0.1",
]
authors = [
    {name = "Tim Felgentreff", email = "timfelgentreff@gmail.com"},
]
maintainers = [
    {name = "Tim Felgentreff", email = "timfelgentreff@gmail.com"},
]

[project.urls]
Repository = "https://github.com/timfel/mpeg1fs.git"
Issues = "https://github.com/timfel/issues"

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
]

[project.scripts]
mpeg1fs = "mpeg1fs:main"

[tool.setuptools_scm]
